Added try/catch for error handling

+++ backend/controllers/todoController.js
@@ -1,5 +1,7 @@
-// IMPORT TODO MODEL
+// IMPORT REQUIRED FILES AND PACKAGES
 const todoModel = require('../models/todoModel');
+const todoErrorHandler = require('../errorHandling/todoErrorHandler');
+const mongoose = require('mongoose');
 
 
 // CREATE METHODS TO PERFORM DIFFERENT ACTIONS
@@ -21,7 +23,8 @@ const createTodos = async (req, res) => {
         res.json({ todo });
     }
     catch(err) {
-        res.json({ error: err.message });
+        const error = todoErrorHandler(err);
+        res.json({ error });
     }
 }
 
@@ -29,7 +32,16 @@ const deleteTodo = async (req, res) => {
     const { id } = req.params;
     
     try {
+        if(!mongoose.Types.ObjectId.isValid(id)) {
+            throw new Error('Not a valid ID!');
+        }
+
         const todo = await todoModel.findByIdAndDelete({ _id: id });
+
+        if(!todo) {
+            throw new Error('Could not find the todo!');
+        }
+
         res.json({ todo });
     }
     catch(err) {
